Refactor -- WeddingGifts앱 관련 View 테스트 test_views.py에 통합.
